requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.2
Django==4.2.4
django-acme==0.2.4
django-cors-headers==4.2.0
django-filter==23.2
django-letsencrypt==4.1.0
django-simpleui==2023.3.1
djangorestframework==3.14.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
mysqlclient==2.2.0
Pillow==9.5.0
pycparser==2.21
pycryptodomex==3.17
PyJWT==2.6.0
pyOpenSSL==23.1.1
pytz==2023.3
redis==4.6.0
requests==2.31.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
